move theme types
