+++ frontend/src/metabase/embedding-sdk/types/theme/private.ts
