+++ frontend/src/metabase/embedding-sdk/types/utils.ts
@@ -1,0 +1,4 @@
+/** Makes every property in the object optional. */
+export type DeepPartial<T> = {
+  [K in keyof T]?: T[K] extends object ? DeepPartial<T[K]> : T[K];
+};
