+++ frontend/src/metabase/visualizations/components/ChartTooltip/EChartsTooltip/EChartsTooltip.stories.tsx
@@ -1,5 +1,5 @@
 import { SdkVisualizationWrapper } from "__support__/storybook";
-import { type MetabaseTheme, defineMetabaseTheme } from "embedding-sdk";
+import type { MetabaseTheme } from "metabase/embedding-sdk/types/theme";
 import { Box, Flex } from "metabase/ui";
 
 import { EChartsTooltip, type EChartsTooltipProps } from "./EChartsTooltip";
@@ -35,7 +35,7 @@ export const LightTheme = {
     </SdkVisualizationWrapper>
   ),
   args: {
-    theme: defineMetabaseTheme({
+    theme: {
       components: {
         tooltip: {
           textColor: "#2f3542",
@@ -44,6 +44,6 @@ export const LightTheme = {
           focusedBackgroundColor: "#f1f2f6",
         },
       },
-    }),
+    },
   },
 };
